% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprof.R
\name{cprof}
\alias{cprof}
\title{Composition Profile}
\usage{
cprof(sequence, ID)
}
\arguments{
\item{sequence}{Character vector of sequences, do not use FASTA format}

\item{ID}{Character vector containing the names of the proteins}
}
\value{

}
\description{
Composition Profile
}
\examples{
\dontrun{
cprof(PROTEINS$SEQUENCE, PROTEINS$ID)
}
}
