% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CH_CDF.R
\name{plot_CH_CDF}
\alias{plot_CH_CDF}
\title{Creat a CH-CDF plot for multiple proteins}
\usage{
plot_CH_CDF(delta_CDF, delta_CH)
}
\arguments{
\item{delta_CDF}{list of delta CDF values}

\item{delta_CH}{List of delta CH values}
}
\value{

}
\description{
Creat a CH-CDF plot for multiple proteins
}
\examples{
\dontrun{
plot_CH_CDF(delta_CDF,delta_CH)
}
}
