% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLXT.R
\name{VLXT}
\alias{VLXT}
\title{Run PONDR VLXT predictor for each protein}
\usage{
VLXT(FASTA, ID)
}
\arguments{
\item{FASTA}{List of FASTA formatted sequences}

\item{ID}{Character vector of protein names}
}
\value{

}
\description{
Run PONDR VLXT predictor for each protein
}
\examples{
\dontrun{
PONDR_VLXT <- VLXT(FASTA,ID)
}

}
